pyproject.toml
[tool.poetry]
name = "ark-sdk-python"
version = "2.0.0"
description='Official Ark SDK / CLI for CyberArk Identity Security Platform'
authors = ["CyberArk <cyberark@cyberark.com>", "Ofir Iluz <ofir.iluz@cyberark.com"]
readme = "README.md"
license = "Apache-2.0"
classifiers = [
    'Intended Audience :: Developers',
    'Natural Language :: English',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12'
]
repository = "https://github.com/cyberark/ark-sdk-python"
packages = [{ include = "ark_sdk_python" }]

[tool.poetry.group.dev.dependencies]
isort = "*"
pylint = "*"
black = "*"
toml = "*"
GitPython = "*"
wheel = "*"
mike = "*"
twine = "*"
mkdocstrings = "*"
mkdocstrings-python = "*"
mkdocs-material = "*"
mkdocs-gen-files = "*"
mkdocs-literate-nav = "*"
mkdocs-section-index = "*"
pymdown-extensions = "*"
pytest = "*"
pytest-mock = "*"
python-dotenv = "*"

[tool.poetry.dependencies]
python = "^3.9"

requests = "*"
keyring = "*"
"keyrings.cryptfile" = "*"
colorama = "*"
pydantic = "2.*"
argcomplete = "*"
pyhumps = "*"
overrides = "*"
packaging = "*"
deepdiff = "7.*.*"
pycryptodome = "*"
dill = "*"
urlextract = "*"
inquirer = "*"
progress = "*"
aenum = "*"
cachetools = "*"
pyjwt = "*"
fake-useragent = "1.5.1"
tzlocal = "*"
pyyaml = "*"
python-dateutil = "*"

[tool.poetry.scripts]
ark = 'ark_sdk_python.ark:main'

[tool.poetry.build]
generate-setup-file = false
script = "build.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
